corporationspecialisations

+++ Core/Model/CorporationSpecialisation.cs
@@ -17,6 +17,6 @@ namespace Core.Model
         public int SpecialisationId { get; set; }
         public Specialisation Specialisation { get; set; }
         [Column("ratio")]
-        public int Ratio { get; set; }
+        public decimal Ratio { get; set; }
     }
 }
